test_requirements.txt
pytest-cov
flake8
flake8-quotes
flake8-comprehensions
flake8-commas
